+++ selene/schema/base.py
@@ -21,7 +21,7 @@
 import graphene
 
 from selene.schema.parser import parse_uuid
-from selene.schema.utils import get_text_from_element
+from selene.schema.utils import get_datetime_from_element, get_text_from_element
 
 
 class NameObjectTypeMixin:
@@ -44,3 +44,35 @@ class BaseObjectType(
     NameObjectTypeMixin, UUIDObjectTypeMixin, graphene.ObjectType
 ):
     """ A base object type resolving an ID and name """
+
+
+class CACertificateMixin:
+
+    md5_fingerprint = graphene.String(
+        description="md5 fingerprint of the CA certificate"
+    )
+
+    issuer = graphene.String(
+        description="Issuer identification of the CA certificate"
+    )
+
+    activation_time = graphene.DateTime(
+        description=(
+            "Datetime when the CA certificate is active and considered valid"
+        )
+    )
+    expiration_time = graphene.DateTime(
+        description="Datetime when the CA certificate will expire"
+    )
+
+    def resolve_md5_fingerprint(root, _info):
+        return get_text_from_element(root, 'md5_fingerprint')
+
+    def resolve_issuer(root, _info):
+        return get_text_from_element(root, 'issuer')
+
+    def resolve_activation_time(root, _info):
+        return get_datetime_from_element(root, 'activation_time')
+
+    def resolve_expiration_time(root, _info):
+        return get_datetime_from_element(root, 'expiration_time')