+++ selene/schema/port_list/fields.py
@@ -37,15 +37,21 @@ from selene.schema.utils import (
 
 
 class PortRangeType(graphene.Enum):
+    """ PortRange type. Either TCP or UDP """
+
     class Meta:
         enum = GvmPortRangeType
 
 
 class PortRange(graphene.ObjectType):
-    uuid = graphene.UUID(name='id')
-    start = graphene.Int()
-    end = graphene.Int()
-    protocol_type = graphene.String()
+    """ A range of ports in a port list """
+
+    uuid = graphene.UUID(name='id', description='ID of the port range')
+    start = graphene.Int(description='Starting port of the range')
+    end = graphene.Int(description='Ending port of the range')
+    port_range_type = graphene.Field(
+        PortRangeType, name='type', description='Type of the port range'
+    )
 
     def resolve_uuid(root, _info):
         return parse_uuid(root.get('id'))
@@ -56,14 +62,22 @@ class PortRange(graphene.ObjectType):
     def resolve_end(root, _info):
         return get_int_from_element(root, 'end')
 
-    def resolve_protocol_type(root, _info):
-        return get_text_from_element(root, 'type')
+    def resolve_port_range_type(root, _info):
+        type_string: str = get_text_from_element(root, 'type')
+        if not type_string:
+            return None
+
+        return type_string.upper()
 
 
 class PortCount(graphene.ObjectType):
-    count_all = graphene.Int(name='all')  # 'all' is a built-in python method
-    tcp = graphene.Int()
-    udp = graphene.Int()
+    """ Aggregation of ports counts """
+
+    count_all = graphene.Int(
+        name='all', description='Total number of ports'
+    )  # 'all' is a built-in python method
+    tcp = graphene.Int(description='Number of TCP ports')
+    udp = graphene.Int(description='Number of UDP ports')
 
     def resolve_count_all(root, _info):
         return get_int_from_element(root, 'all')
@@ -76,20 +90,26 @@ class PortCount(graphene.ObjectType):
 
 
 class PortListTarget(BaseObjectType):
-    pass
+    """ A target referenced by a PortList """
 
 
 class PortList(EntityObjectType):
+    """ A list of ports to scan """
+
     class Meta:
         default_resolver = find_resolver
 
-    port_ranges = graphene.List(PortRange)
-    port_count = graphene.Field(PortCount)
-    targets = graphene.List(PortListTarget)
+    port_ranges = graphene.List(
+        PortRange, description="Port ranges in this port list"
+    )
+    port_count = graphene.Field(PortCount, description="Summary of ports")
+    targets = graphene.List(
+        PortListTarget, description="Targets using this port list"
+    )
 
     def resolve_port_ranges(root, _info):
         port_ranges = root.find('port_ranges')
-        if len(port_ranges) == 0:
+        if port_ranges is None or len(port_ranges) == 0:
             return None
         return port_ranges.findall('port_range')